+++ lib/utils/parse-data-default.ts
@@ -1,24 +1,12 @@
+import { stripQuotes } from './strip-quotes';
+
 /**
- * Strip leading/trailing quotes from a string and handle null values.
+ * Handle Oracle/SQLite returning unquoted string null to indicate null default.
  */
 export function parseDataDefault(value: string | null): string | null {
-  if (value === null) {
+  if (value?.trim().toLocaleLowerCase() === 'null') {
     return null;
   }
 
-  const trimmed = value.trim();
-
-  // Oracle/SQLite can return unquoted string null to indicate null default
-  if (trimmed.toLocaleLowerCase() === 'null') {
-    return null;
-  }
-
-  if (
-    (trimmed.startsWith(`'`) && trimmed.endsWith(`'`)) ||
-    (trimmed.startsWith('"') && trimmed.endsWith('"'))
-  ) {
-    return trimmed.slice(1, -1);
-  }
-
-  return value;
+  return stripQuotes(value);
 }
